Update WritingToolApp.py
Custom Prompt often responded with ERROR_TEXT_INCOMPATIBLE_WITH_REQUEST. 
To prevent that, I did some test and reformed the System Prompt to exclude the request to put out the ERROR Message if the text is not compatible. 
I also updated the User-Prompt template to bring better results for text and code. The user request is described as <task> and the marked text is the <context> for more generality.